==> Studies/S2-surveys/2_surveys/3_VIS25_submission/0.0_SUB1006_S1_2_wrangling.Rmd <==
---
title: "S1-2 DATA WRANGLING—SUBMISSION 1006"
author: "ANONYMIZED"
date: "2025-02-24"
output:
  html_document:
    theme: flatly
    code_folding: hide
    fig_caption: yes
    number_sections: yes
    toc: yes
    toc_float:
      collapsed: yes
      smooth_scroll: yes
    toc_depth: 5
  pdf_document:
    toc: yes
    toc_depth: '4'
always_allow_html: yes
font-family: DejaVu Sans
mainfont: DejaVu Sans
---

```{r setup, include=FALSE, message=FALSE}
knitr::opts_chunk$set(echo = TRUE)

#UTILITIES
library(Hmisc) # %nin% operator
library(psych) #describe()
library(tidyverse) #all the things
library(magrittr) #special pipes like %<>%
library(lubridate) #dealing with dates
library(kableExtra) #printing tables

#EXPLORATORY DATA ANALYSIS
library(summarytools) #data quality
library(qacBase) #EDA

#CONFIG
options(readr.show_col_types = FALSE) #don't show coltypes on read_csv
n_blocks = 6 #number of stimulus blocks in study design

```

*This notebook contains code to wrangle raw data from multiple qualtrics surveys into a single normalized dataset of (valid) participant responses.*

The following dataframes are created:

-   `df_data` represents **response data** for all valid participants (wide format) — answers are not stored as factors; not useful for numeric analysis but just as a record of participant-level responses

-   `df_participants` represents **demographic data** for all valid participants

-   `df_qda_wide` includes all valid responses in wide format, to be used for free response data analysis (wide)

-   `df_qda_long` represents response and demographic data for all valid participants (used to generate a spreadsheet for analysis for free response data) *at the trial (i.e. stimulus) level* (i.e. long).

-   `df_questions` include valid responses at the most granular level, where each row refers to a participant+stimulus+question (long format) [318p X 5s X 28q]

-   `df_sd_questions_long` includes only the semantic differential-type questions (long)

-   `df_sd_questions_wide` is a wide version of the `df_sd_questions` dataframe useful for some graphs (318p X 11q = 3498 rows)

-   `df_tools` contains responses to the tool id and tool confidence questions, where tool_id is a multi-select Q

-   `df_actions` contains responses to the action id and action confidence questions, where df_action is a multi-select Q

-   `df_graphs_full` is a stimulus (ie. graph image) level dataframe where each row contains columns for a participant and all the Qs for that graph

-   `df_graphs` is df_graphs_full w/o the free response and multiselect Qs

# STANDARDIZE

This notebook takes raw data files from (3) Qualtrics surveys (Study 1, Study 2, Study 3), standardizing their format for further data analysis. Note that all raw data comes from Qualtrics, and each Qualtrics survey includes a different stimulus block (set of 5 stimuli).

1.  **Study 1:** We collected data sampling directly from the Tumblr social media platform produces files with slightly different metadata (several fields required for tracking Prolific are not included). The Qualtrics surveys for Tumblr used randomization logic to assign each Tumblr participant to one of the six stimulus blocks. **(All stimulus blocks, n \~ 12-15 / block)**

2.  **Study 0:** Initial data collection via PROFLIFIC occurred via Qualtrics surveys named `MAG_S2_PROLIFIC-DATACOLLAR_0` and `MAG_S2_PROLIFIC-DATACOLLAR_0`, each presenting participants with one common graph ("block 0", `STIMULUS = B0-0`) and four subsequent graphs in random order ("block 1"). These data were collected via Prolific, with different demographic restrictions (i.e. 'data collar' and 'blue collar' made use of different values in the Prolific participant background survey on the 'technology use at work' and 'job title' questions.) However, when we duplicated these questions in our own demographic surveys, we found participants' answers often did not match their values on the Prolific background survey, and thus we abandoned this stratified sampling strategy and continued data collection with no demographic restrictions. **(Stimulus block #1 only; n = 40 (20 blue-collar; 20 data-collar) )**

3.  **Study 2:** General (non-stratified) sampling via PROLIFIC collecting data for all 6 stimulus blocks. **(n =40/block )**

4.  *Data for Study 3 (a repeated measures design where participants rate the stimuli with obscured and unobscured versions of each image)* This wrangling is handled in a separate notebook.

## Import Configuration Files

*Import files with reference information for stimuli `ref_stimuli` and qualtrics surveys `ref_surveys`.*

```{r setup-references, message=FALSE, warning = FALSE}

############## IMPORT STIMULI FILE
#contains data for each stimulus used in study 2
ref_stimuli <- read_csv("data/input/REFERENCE/stimuli.csv", col_names = TRUE) %>% 
  mutate(
    BLOCK = as.factor(BLOCK), 
    STIMULUS_CATEGORY = as.factor(CATEGORY),
    ID = as.factor(ID),
    MAKER_ID = as.factor(MAKER_ID)) %>% 
  select(-CATEGORY)
saveRDS(ref_stimuli, file = "data/input/REFERENCE/ref_stimuli.rds")


############## STUDY ID FILE
#most blocks were run as separate qualtrics surveys with diffferent recruitments in Prolific
#Tumblr was run with all blocks and randomization
ref_surveys <- read_csv("data/input/REFERENCE/studies.csv", col_names = TRUE) %>%
  mutate(
    ID.Study = as.factor(ID.Study),
    Assigned.Block = as.factor(Assigned.Block),
    Distribution = as.factor(Distribution),
    Prolific.Name = as.factor(Prolific.Name),
    Qualtrics.URL = as.factor(Qualtrics.URL),
    Qualtrics.Survey = as.factor(Qualtrics.Survey),
    Sample = as.factor(Sample),
    Study = as.factor(Study)
)
saveRDS(ref_surveys, file = "data/input/REFERENCE/ref_surveys.rds")

```

## Define Labels

*Define labels used in survey questions and response options; used to generate labels for graphs and filtering of datasets in analysis files.*

```{r setup-labels}

############## BUILD LABELS
ref_stim_id <- levels(ref_stimuli$ID)
ref_blocks <- c("block1", "block2", "block3", "block4", "block5", "block6")
ref_cat_questions <- c("MAKER_ID","MAKER_AGE","MAKER_GENDER")
ref_free_response <- c("MAKER_DETAIL", "MAKER_EXPLAIN", "TOOL_DETAIL", "CHART_EXPLAIN")
ref_conf_questions <- c("MAKER_CONF", "AGE_CONF", "GENDER_CONF", "TOOL_CONF")


############## LABELS FOR RAW VALUE OF SD QUESTIONS
ref_sd_questions <- c("MAKER_DESIGN","MAKER_DATA",
                      "MAKER_POLITIC", "MAKER_ARGUE","MAKER_SELF","MAKER_ALIGN","MAKER_TRUST",
                      "CHART_TRUST", "CHART_INTENT", "CHART_LIKE", "CHART_BEAUTY")

left <- c("professional","professional",
          "left-leaning","confrontational","altruistic","does NOT share","untrustworthy",
          "untrustworthy", "inform", "NOT at all","NOT at all")

right <- c("layperson","layperson", 
           "right-leaning","diplomatic", "selfish", "DOES share", "trustworthy",
           "trusthworthy", "persuade","very much",  "very much")

category <- c("COMPETENCY","COMPETENCY",
              "MAKER","MAKER","MAKER","MAKER","MAKER",
              "CHART","CHART","CHART","CHART")
              

ref_labels <- as.data.frame(cbind(ref_sd_questions, left,right,category))
rownames(ref_labels) <- ref_sd_questions

write.csv(ref_labels, file = "data/input/REFERENCE/ref_labels.csv", na="")
saveRDS(ref_labels, file = "data/input/REFERENCE/ref_labels.rds")



############## LABELS FOR Z-scored VALUE OF SD QUESTIONS
ref_sd_questions_z <- c("MAKER_DESIGN_z","MAKER_DATA_z",
                      "MAKER_POLITIC_z", "MAKER_ARGUE_z","MAKER_SELF_z","MAKER_ALIGN_z","MAKER_TRUST_z",
                      "CHART_TRUST_z", "CHART_INTENT_z", "CHART_LIKE_z", "CHART_BEAUTY_z")




############## LABELS FOR ABS VALUE OF SD QUESTIONS
ref_sd_questions_abs <- c("MAKER_DESIGN","MAKER_DATA",
                      "MAKER_POLITIC", "MAKER_ARGUE","MAKER_SELF","MAKER_ALIGN","MAKER_TRUST",
                      "CHART_TRUST", "CHART_INTENT", "CHART_LIKE", "CHART_BEAUTY")

left <- c("(neutral)","(neutral)",
          "(neutral)","(neutral)","(neutral)","(neutral)","(neutral)",
          "(neutral)","(neutral)", "(neutral)", "(neutral)")

right <- c("professional/layperson","professiona/layperson",
           "left-leaning/right-leaning","confrontational/diplomatic", "altruistic/selfish","does NOT share/DOES share","untrustworthy/trustworthy",
            "untrustworthy/trustworthy","inform/persuade", "NOT at all/VERY much","NOT at all/VERY much" )

category <- c("COMPETENCY","COMPETENCY",
              "MAKER","MAKER","MAKER","MAKER","MAKER",
              "CHART","CHART","CHART","CHART")

ref_labels_abs <- as.data.frame(cbind(ref_sd_questions_abs,left,right,category))
rownames(ref_labels_abs) <- ref_sd_questions_abs

write.csv(ref_labels_abs, file = "data/input/REFERENCE/ref_labels_abs.csv", na="")
saveRDS(ref_labels_abs, file = "data/input/REFERENCE/ref_labels_abs.rds")


```

## Define Factors

*Define order of factors for multiple-choice variables*

```{r define-factor-orders}

order_study = c("Study1","Study2","Study3","Study0") ##NEW for VIS
order_sample = c("TUMBLR","DATACOLLAR","BLUECOLLAR", "GENERAL") ## NEW for VIS
order_distribution = c("TUMBLR","PROLIFIC")
order_maker = c("individual","organization","education","business","news","political")
order_age = c("boomer", "gen-x","millennial","gen-z")
order_gender = c("Other","Female","Male" )
order_encounter = c("scroll", "engage")
order_tools= c("?","design_basic","design_advanced","viz_basic","viz_advanced","programming")
order_actions = c("nothing","unfollow/block","comment","share","share w/ comment","seek information")
## condensed categories for actions 
order_actions4 = c("nothing","unfollow/block","comment/share","seek information")
order_actions3 = c("nothing","solo engagement","social engagement")
order_actions2 = c("nothing","something")

```

## Import Raw Survey Files

*Here we import raw qualtrics data files for each survey, and normalize column names to produce a single raw data file including all attempts at completing all the surveys: `df_raw`. This intermediary file is not used for analysis, but rather for further wrangling into participant, trial, and question level files*

```{r WRANGLE-INPUT, message=FALSE, warning=FALSE}

#### IMPORT RAW DATA FILES ########################################################
# will always be the unaltered version of imported data
# 1 row per subject 
df_raw_datacollar <- read_csv("data/input/CLEAN/CLEAN_MAG_S2_PROLIFIC_datacollarpilot_B1.csv", col_names = TRUE)
df_raw_bluecollar <- read_csv("data/input/CLEAN/CLEAN_MAG_S2_PROLIFIC_bluecollarpilot_B1.csv", col_names = TRUE)
df_raw_b1 <- read_csv("data/input/CLEAN/CLEAN_MAG_S2_PROLIFIC_GENERAL_B1.csv", col_names = TRUE)
df_raw_b2 <- read_csv("data/input/CLEAN/CLEAN_MAG_S2_PROLIFIC_GENERAL_B2.csv", col_names = TRUE)
df_raw_b3 <- read_csv("data/input/CLEAN/CLEAN_MAG_S2_PROLIFIC_GENERAL_B3.csv", col_names = TRUE)
df_raw_b4 <- read_csv("data/input/CLEAN/CLEAN_MAG_S2_PROLIFIC_GENERAL_B4.csv", col_names = TRUE)
df_raw_b5 <- read_csv("data/input/CLEAN/CLEAN_MAG_S2_PROLIFIC_GENERAL_B5.csv", col_names = TRUE)
df_raw_b6 <- read_csv("data/input/CLEAN/CLEAN_MAG_S2_PROLIFIC_GENERAL_B6.csv", col_names = TRUE)
df_raw_tumblr_paid <- read_csv("data/input/CLEAN/CLEAN_MAG_S2_TUMBLR_PAID_ALL.csv", col_names = TRUE)
df_raw_tumblr_free <- read_csv("data/input/CLEAN/CLEAN_MAG_S2_TUMBLR_FREE_ALL.csv", col_names = TRUE)

# drop first two rows (these are qualtrics_specs rows)
df_raw_datacollar <- df_raw_datacollar[-c(1:2),]
df_raw_bluecollar <- df_raw_bluecollar[-c(1:2),]
df_raw_b1 <- df_raw_b1[-c(1:2),]
df_raw_b2 <- df_raw_b2[-c(1:2),]
df_raw_b3 <- df_raw_b3[-c(1:2),]
df_raw_b4 <- df_raw_b4[-c(1:2),]
df_raw_b5 <- df_raw_b5[-c(1:2),]
df_raw_b6 <- df_raw_b6[-c(1:2),]
df_raw_tumblr_paid <- df_raw_tumblr_paid[-c(1:2),]
df_raw_tumblr_free <- df_raw_tumblr_free[-c(1:2),]

# ADD DUMMY COLS TO DATA/BLUE COLLAR PILOT DATA 
# necessary b/c pilot (block 1) did not have chart_action behavioural question
# x <- compare_df_cols(df_raw_b2, df_raw_pilot)
df_raw_pilot <- rbind(df_raw_datacollar, df_raw_bluecollar) %>% 
  mutate(
    '0_Q_B0_CHART_ACTION' = NA,
    '1_Q_B1_CHART_ACTION' = NA,
    '1_Q_B2_CHART_ACTION' = NA,
    '1_Q_B3_CHART_ACTION' = NA,
    '1_Q_B4_CHART_ACTION' = NA,
    '1_Q_B5_CHART_ACTION' = NA,
    '1_Q_B6_CHART_ACTION' = NA,
    '2_Q_B1_CHART_ACTION' = NA,
    '2_Q_B2_CHART_ACTION' = NA,
    '2_Q_B3_CHART_ACTION' = NA,
    '2_Q_B4_CHART_ACTION' = NA,
    '2_Q_B5_CHART_ACTION' = NA,
    '2_Q_B6_CHART_ACTION' = NA,
    '3_Q_B1_CHART_ACTION' = NA,
    '3_Q_B2_CHART_ACTION' = NA,
    '3_Q_B3_CHART_ACTION' = NA,
    '3_Q_B4_CHART_ACTION' = NA,
    '3_Q_B5_CHART_ACTION' = NA,
    '3_Q_B6_CHART_ACTION' = NA,
    '4_Q_B1_CHART_ACTION' = NA,
    '4_Q_B2_CHART_ACTION' = NA,
    '4_Q_B3_CHART_ACTION' = NA,
    '4_Q_B4_CHART_ACTION' = NA,
    '4_Q_B5_CHART_ACTION' = NA,
    '4_Q_B6_CHART_ACTION' = NA
  )

# BIND COLUMNS RAW prolific datasets 
df_raw_prolific <- rbind(df_raw_pilot, df_raw_b1, df_raw_b2, df_raw_b3, df_raw_b4, df_raw_b5, df_raw_b6)

# RETROFIT SOME COLNAMES FOR COMPATIBILITY WITH TUMBLR 
df_raw_prolific <- df_raw_prolific %>% 
  #drop T_BROWSER cols [these are blank]
  select(-contains("T_BROWSER"), -T_EMAIL) %>% 
  mutate(
    Q_RelevantIDDuplicate = NA, 
    Q_RelevantIDDuplicateScore = NA, 
    Q_RelevantIDFraudScore = NA, 
    Q_RelevantIDLastStartDate = NA, 
    RANDOM_BLOCK = NA
  ) %>% rename_with(
    stringr::str_replace, 
      pattern = "P_BROWSER", replacement = "BROWSER",
  ) %>% 
  dplyr::rename(
    BROWSER_OS = `BROWSER_Operating System`
  ) %>% mutate_all(funs(str_replace(., "millenial", "millennial")))


# BIND COLUMNS RAW prolific datasets 
df_raw_tumblr <- rbind(df_raw_tumblr_paid, df_raw_tumblr_free)

# RETROFIT SOME TUMBLR COLNAMES FOR MERGING WITH PROLIFIC 
df_raw_tumblr <- df_raw_tumblr %>% 
  select( - t_example_question1, -t_example_question2, -D_email) %>% 
  mutate(
    #add empty browser cols; didn't collect these in tumblr qualtrics
    BROWSER_Browser = NA,
    BROWSER_OS = NA,
    BROWSER_Resolution = NA,
    BROWSER_Version = NA, 
    ID_PROLIFIC = "TUMBLR",
    ID_SESSION = "TUMBLR",
    ID_STUDY = "TUMBLR"
  ) %>% mutate_all(funs(str_replace(., "millenial", "millennial")))

#dataframe column comparisons 
# x <- janitor::compare_df_cols(df_raw_prolific, df_raw_tumblr)

# JOINT PROLIFIC AND TUMBLR DATAFRAMES
df_raw <- rbind(df_raw_prolific, df_raw_tumblr)

df_raw <- df_raw %>% 
  #reorder 
  select( RANDOM_BLOCK, 
          End_State,
          StartDate:randomize_common, 
          ID_PROLIFIC:ID_SESSION, 
          D_gender:FEEDBACK, 
          PROLIFIC_PID: FL_14_DO,
          Q_RelevantIDDuplicate: Q_RelevantIDLastStartDate, 
          `0_Q_B0_ENCOUNTER`: `0_Q_B0_CHART_LATENCY_Click Count`, 
          `0_Q_B0_CHART_ACTION`, 
          `1_Q_B1_loop-number` : `1_Q_B1_CHART_LATENCY_Click Count`,
          `1_Q_B1_CHART_ACTION`, 
          `2_Q_B1_loop-number` : `2_Q_B1_CHART_LATENCY_Click Count`,
          `2_Q_B1_CHART_ACTION`, 
          `3_Q_B1_loop-number` : `3_Q_B1_CHART_LATENCY_Click Count`,
          `3_Q_B1_CHART_ACTION`, 
          `4_Q_B1_loop-number` : `4_Q_B1_CHART_LATENCY_Click Count`,
          `4_Q_B1_CHART_ACTION`, 
          `1_Q_B2_loop-number` : `1_Q_B2_CHART_LATENCY_Click Count`,
          `1_Q_B2_CHART_ACTION`, 
          `2_Q_B2_loop-number` : `2_Q_B2_CHART_LATENCY_Click Count`,
          `2_Q_B2_CHART_ACTION`, 
          `3_Q_B2_loop-number` : `3_Q_B2_CHART_LATENCY_Click Count`,
          `3_Q_B2_CHART_ACTION`, 
          `4_Q_B2_loop-number` : `4_Q_B2_CHART_LATENCY_Click Count`,
          `4_Q_B2_CHART_ACTION`, 
          `1_Q_B3_loop-number` : `1_Q_B3_CHART_LATENCY_Click Count`,
          `1_Q_B3_CHART_ACTION`, 
          `2_Q_B3_loop-number` : `2_Q_B3_CHART_LATENCY_Click Count`,
          `2_Q_B3_CHART_ACTION`, 
          `3_Q_B3_loop-number` : `3_Q_B3_CHART_LATENCY_Click Count`,
          `3_Q_B3_CHART_ACTION`, 
          `4_Q_B3_loop-number` : `4_Q_B3_CHART_LATENCY_Click Count`,
          `4_Q_B3_CHART_ACTION`, 
          `1_Q_B4_loop-number` : `1_Q_B4_CHART_LATENCY_Click Count`,
          `1_Q_B4_CHART_ACTION`, 
          `2_Q_B4_loop-number` : `2_Q_B4_CHART_LATENCY_Click Count`,
          `2_Q_B4_CHART_ACTION`, 
          `3_Q_B4_loop-number` : `3_Q_B4_CHART_LATENCY_Click Count`,
          `3_Q_B4_CHART_ACTION`, 
          `4_Q_B4_loop-number` : `4_Q_B4_CHART_LATENCY_Click Count`,
          `4_Q_B4_CHART_ACTION`, 
          `1_Q_B5_loop-number` : `1_Q_B5_CHART_LATENCY_Click Count`,
          `1_Q_B5_CHART_ACTION`, 
          `2_Q_B5_loop-number` : `2_Q_B5_CHART_LATENCY_Click Count`,
          `2_Q_B5_CHART_ACTION`, 
          `3_Q_B5_loop-number` : `3_Q_B5_CHART_LATENCY_Click Count`,
          `3_Q_B5_CHART_ACTION`, 
          `4_Q_B5_loop-number` : `4_Q_B5_CHART_LATENCY_Click Count`,
          `4_Q_B5_CHART_ACTION`, 
          `1_Q_B6_loop-number` : `1_Q_B6_CHART_LATENCY_Click Count`,
          `1_Q_B6_CHART_ACTION`, 
          `2_Q_B6_loop-number` : `2_Q_B6_CHART_LATENCY_Click Count`,
          `2_Q_B6_CHART_ACTION`, 
          `3_Q_B6_loop-number` : `3_Q_B6_CHART_LATENCY_Click Count`,
          `3_Q_B6_CHART_ACTION`, 
          `4_Q_B6_loop-number` : `4_Q_B6_CHART_LATENCY_Click Count`,
          `4_Q_B6_CHART_ACTION`
          #df_raw 936 vars           
  ) 



#DROP WIP DATAFRAMES 
rm(df_raw_datacollar, df_raw_bluecollar, df_raw_b1, df_raw_b2, df_raw_b3, df_raw_b4, df_raw_b5, df_raw_b6, df_raw_pilot, df_raw_prolific, df_raw_tumblr, df_raw_tumblr_free, df_raw_tumblr_paid)

```

## Wrangle Standardized File

Here we drop irrelevant columns (qualtrics junk) and set factors. The resulting `df_data` frame is wide (1 row/participant, but includes redundant columns for the blocks they were not assigned to).

```{r WRANGLE-FACTORS, warning=FALSE}

####WRANGLE MASTER PARTICIPANT-LEVEL DF  #####################################################
#### MASTER WIDE FORMAT DATA FRAME [1 row / valid qualtrics submission] ######################

##RENAME AND SET CORRECT DATA TYPES
df_data <- df_raw %>% 
  select(
    -EndDate, -IPAddress, -RecordedDate,
    -RecipientLastName, -RecipientFirstName, -RecipientEmail,
    -ExternalReference, -LocationLatitude, -LocationLongitude, 
    -DistributionChannel, -UserLanguage, -Q_RecaptchaScore,
    -BROWSER_Version, -BROWSER_Resolution, 
    -CONSENT, -ELIGIBILITY, 
    -randomize_common,
    #hidden q that controls common stimulus url
    -stimulus_common, 
    #not actually randomization order
    -FL_14_DO, 
    -contains("First Click"), -contains("Last Click"), -contains("Click Count"),
    -D_politicalParty_DO, 
    -ID_PROLIFIC, -ID_STUDY, -ID_SESSION #redundant to other cols 
  ) %>% 
  dplyr::rename(
    duration.sec = `Duration (in seconds)`,
    EndState = End_State, 
    TerminateFlag = Q_TerminateFlag,
    Source = Status, #where the survey originated from (should not be preview or test)
    PLATFORM = Q_PLATFORM,
    ID.Qualtrics = ResponseId,
    ID.Prolific = PROLIFIC_PID,
    ID.Study = STUDY_ID,
    ID.Session = SESSION_ID,
    # P_BROWSER_OS = `P_BROWSER_Operating System`,
    # T_BROWSER_OS = `T_BROWSER_Operating System`,
    SCREEN_workFunction_TEXT = SCREEN_workFunction_22_TEXT, 
    SCREEN_socialMedia_TEXT = SCREEN_socialMedia_18_TEXT,
    D_politicalParty_OTHER = D_politicalParty_4_TEXT,
    D_politicsSocial = D_politicsSocial_1,
    D_politicsFiscal = D_politicsFiscal_2
 ) %>% 
  mutate(
    #SET FACTORS 
    D_politicsSocial = as.numeric(D_politicsSocial),
    D_politicsFiscal = as.numeric(D_politicsFiscal),
    ID.Study = factor(ID.Study),
    ID.Qualtrics = factor(ID.Qualtrics),
    ID.Prolific = factor(ID.Prolific),
    ID.Session = factor(ID.Session),
    PLATFORM = factor(PLATFORM),
    Source = factor(Source),
    Finished = as.logical(Finished),
    TerminateFlag = factor(TerminateFlag),
    EndState = factor(EndState), 
    D_gender = factor(D_gender), 
    D_gender_collapsed = fct_collapse(D_gender, 
                                male = "Male", 
                                female = "Female", 
                                other = c("Non-binary / third gender", "Prefer not to say", "Prefer to self-describe")),
    D_age = factor(D_age), 
    D_income = factor(D_income, 
                      levels = c(
                        "Prefer not to say",
                        "Less than $25,000",
                        "$25,000-$49,999"  ,
                        "$50,000-$74,999"  ,
                        "$75,000-$99,999"  ,
                        "$100,000-$149,999",
                        "$150,000 or more" 
                      )),
    D_employmentStatus = factor(D_employmentStatus),
    duration.sec = as.numeric(duration.sec), #weird booleans should only be for the test generator
    duration.min = round(duration.sec/60,2),
    Progress = as.numeric(Progress), 
    D_ed = factor(D_education),
    D_education = forcats::fct_na_value_to_level( D_education, level="NA"),
    D_education = forcats::fct_collapse( D_education,
                                no_data = "NA",
                                less_high_school = c("Some high school or less"),
                                high_school = c("High school diploma or GED"),
                                some_college = c("Some college, but no degree", "Some college, no degree"),
                                associates =c( "Associates or technical degree"),
                                undergrad = c("Bachelor’s degree","Bachelor's degree"),
                                grad = c("Graduate or professional degree (MA, MS, MBA, PhD, JD, MD, DDS etc.)",
                                         "Graduate or professional degree (MA, MS, MBA, PhD, JD, MD, DDS, etc)")
    ),
    D_education = factor(D_education,
                                       levels = c("no_data", "less_high_school","high_school", "some_college",
                                         "associates", "undergrad", "grad"),
                                       labels = c("NA", "some high school or less","high school diploma or GED ",
                                                  "some college", "associates or technical degree",
                                                  "undergradudate degree", "graduate or professional degree"),
                                       ),
    D_education_collapsed = forcats::fct_na_value_to_level( D_education, level="NA"),
    D_education_collapsed = forcats::fct_collapse(D_education_collapsed,
                                         hs_or_less = c("NA","some high school or less","high school diploma or GED "),
                                         ugrad = c("some college", "associates or technical degree",
                                                  "undergradudate degree"),
                                         grad = c("graduate or professional degree")
                                         ),
    D_politicalParty = factor(D_politicalParty, levels = c("Democrat", "Other", "No preference", "Independent", "Republican")),
    D_age = factor(D_age, 
                   levels = c("18-24 years old" ,
                              "25-34 years old" ,
                              "35-44 years old" ,
                              "45-54 years old" ,
                              "55-64 years old" ,
                              "65+ years old"   ), 
                   labels = c("18-24", "25-34","35-44","45-54","55-64","65+ years"))
  ) %>%
  #REPLACE RANDOM TRAILING _1 AND _65 FROM QUALTRICS
  rename_with( .cols = contains('_65'), .fn = ~str_replace(.,  pattern = '_65', replacement = '')) %>% 
  rename_with( .cols = contains('_1'), .fn = ~str_replace(.,  pattern = '_1', replacement = '')) %>% 
  #RM _PAGE Submit from LATENCY
  rename_with( .cols = contains('_Page Submit'), .fn = ~str_replace(.,  pattern = '_Page Submit', replacement = '')) %>% 
  #CHANGE DETAIL QUESTION DELIMITER FOR PIVOT PURPOSES 
  rename_with( .cols = contains('_CHART_'), .fn = ~str_replace(.,  pattern = '_CHART_', replacement = '_CHART-')) %>% 
  #CHANGE DETAIL QUESTION DELIMITER FOR PIVOT PURPOSES 
  rename_with( .cols = contains('_MAKER_'), .fn = ~str_replace(.,  pattern = '_MAKER_', replacement = '_MAKER-')) %>% 
  #CHANGE DETAIL QUESTION DELIMITER FOR PIVOT PURPOSES 
  rename_with( .cols = contains('_AGE_'), .fn = ~str_replace(.,  pattern = '_AGE_', replacement = '_AGE-')) %>% 
  #CHANGE DETAIL QUESTION DELIMITER FOR PIVOT PURPOSES 
  rename_with (.cols = contains('_GENDER_'), .fn = ~str_replace(., pattern = '_GENDER_', replacement = '_GENDER-')) %>% 
  #CHANGE DETAIL QUESTION DELIMITER FOR PIVOT PURPOSES 
  rename_with (.cols = contains('_TOOL_'), .fn = ~str_replace(., pattern = '_TOOL_', replacement = '_TOOL-')) %>% 
  select(
    #reordering
    RANDOM_BLOCK:Progress, 
    Finished, EndState, TerminateFlag, 
    Q_RelevantIDDuplicate:Q_RelevantIDLastStartDate,
    ID.Qualtrics, 
    ID.Prolific : ID.Session, 
    duration.sec, duration.min,
    BROWSER_Browser : BROWSER_OS, 
    D_gender_collapsed, D_gender, D_gender_4_TEXT,
    D_race,D_education_collapsed, D_education,
    D_employmentStatus:D_politicsFiscal,
    SCREEN_workMethod: SCREEN_socialMedia_TEXT, 
    PURPOSE, FEEDBACK, PLATFORM,
    `0_Q_B0_ENCOUNTER`: `4_Q_B6_CHART-ACTION`
  ) 

##### JOIN SURVEY-LEVEL DATA 
df_data <- dplyr::left_join(df_data, ref_surveys, by="ID.Study") 

#SET IDS AND ASSIGNED BLOCK TO HANDLE PROLIFIC AND TUMBLR 
df_data <- df_data %>% 
  mutate(
    # SET ASSIGNMENT BLOCK FOR TUMBLR
    Assigned.Block = if_else( (Distribution =="TUMBLR"), RANDOM_BLOCK, Assigned.Block),
    Assigned.Block = factor(Assigned.Block), 
    # PID = if_else( (Distribution =="TUMBLR"), ID.Qualtrics, ID.Prolific),
    PID = factor(ID.Qualtrics),
    Distribution = factor(Distribution, levels=order_distribution), ## ADDED FACTOR ORDER TO KEEP TUBMLR AS FIRST 
    Sample = fct_recode(Sample, TUMBLR= "tumblr", BLUECOLLAR = "blue-collar", DATACOLLAR = "data-collar", GENERAL="general-prolific"), #recode factor levels
    Sample = factor(Sample, levels= order_sample), #set level order
    Study = factor(Study, levels = order_study) #set level order
  ) %>% #DROP RANDOM_BLOC COLUMNS
  select (-RANDOM_BLOCK) %>% 
  select(
    #REORDER
    PID, 
    Distribution, 
    Assigned.Block,
    ID.Qualtrics:ID.Session,
    EndState,
    StartDate: Q_RelevantIDLastStartDate, 
    Prolific.Name, Qualtrics.Survey,  Qualtrics.URL, Description, Sample, Study, 
    duration.sec:PLATFORM,
    `0_Q_B0_ENCOUNTER` : `4_Q_B6_CHART-ACTION`
  )

#DROP WIP DATAFRAMES 
rm(df_raw)
```

# CLEANING

*Here we apply exclusion criteria in order to reduce the raw participant data in `df_data` to a dataframe of valid participant data `df_subjects`. We remove unneeded columns (i.e. qualtrics configuration columns, etc) and transform columns to the the appropriate data types for further analysis.*

Participants were *excluded* from the sample for the following reasons:

-   **abandoned**; Finished = False and Progress \< 100%
-   **fails consent or pre-screening**, Finished = TRUE, Progress = 100%, Q_TerminateFlag = "Screened"
-   **didnot-follow-instructions** free responses indicate they misinterpred the question (most commonly they describe the graph rather than their reactions/judgements/impressions of it)
-   **illegible-english** free responses text is largely illegible

A dataframe `df_exclude` is saved, including **response data** for participants who did not meet inclusion criteria (wide format)

```{r CLEAN-MASTER}

#1B CLEAN MASTER PARTICIPANT-LEVEL DF  #########################################
#### SEGREGATE PARTICIPANTS WHO DID NOT COMPLETE ###############################
## [1 row / qualtrics submission] ################
## NOTE it is common for prolific participants to fail the 
## screening verification, and then try again but change their 
## screening verification answers (ie. one prolific ID for multiple qualrics IDs)
df_exclude <- df_data %>% 
  filter( 
    !is.na(TerminateFlag) | Finished == FALSE | EndState != "COMPLETE"  
  ) %>% 
  select(
    PID, Distribution, ID.Qualtrics, ID.Prolific, ID.Study, Assigned.Block, Study, Source, Progress, 
    Finished, TerminateFlag, EndState,StartDate, duration.min, 
    D_gender:D_politicsFiscal, SCREEN_workMethod:FEEDBACK, Prolific.Name:Study
  )  %>% 
  mutate(
    EndState = if_else( (Progress < 100), "abandoned", EndState),
    EndState = if_else( (str_detect(EndState,"screen")), "screened", EndState),
    EndState = if_else( (TerminateFlag == "Screened" & is.na(EndState)), "screened", EndState),
    EndState = factor(EndState)
  )
write.csv(df_exclude, file = "data/output/Study_1_2/df_exclude.csv", na="")
saveRDS(df_exclude, file = "data/output/Study_1_2/df_exclude.rds")



#### MASTER VALID DATA [WIDE] 1 row per qualtrics entry #########################
## [1 row / qualtrics submission] ################
df_data <- df_data %>% filter(
  PID %nin% df_exclude$PID
  # (Finished == TRUE ) & is.na(TerminateFlag)
) %>% mutate(
  EndState = droplevels(EndState),
  ID.Prolific = droplevels(ID.Prolific),
  ID.Qualtrics = droplevels(ID.Qualtrics), 
  PID = droplevels((PID))
)


##SANITY CHECKS ON EXLUSIONS
#sanity check === SHOULD BE O
#no qualtrics surveys in good data that weren't finished
print("Number of PID entries in df_data AND nofinish/excluded? [should be 0]")
sum(df_data$PID %in% df_exclude$PID)

#sanity check === SHOULD BE O
#no duplicated PROLIFIC ids in good data 
print("Number of PIDs duplicated in df_data print [should be 0]")
sum(duplicated(df_data$PID))
```

# WRANGLE

*Reshape data into files for analysis*

The following dataframes are created:

-   `df_data` represents **response data** for all valid participants (wide format) — answers are not stored as factors; not useful for numeric analysis but just as a record of participant-level responses

-   `df_participants` represents **demographic data** for all valid participants

-   `df_qda_wide` includes all valid responses in wide format, to be used for free response data analysis (wide)

-   `df_qda_long` represents response and demographic data for all valid participants (used to generate a spreadsheet for analysis for free response data) *at the trial (i.e. stimulus) level* (i.e. long).

-   `df_questions` include valid responses at the most granular level, where each row refers to a participant+stimulus+question (long format) [318p X 5s X 28q]

-   `df_sd_questions_long` includes only the semantic differential-type questions (long)

-   `df_sd_questions_wide` is a wide version of the `df_sd_questions` dataframe useful for some graphs (318p X 11q = 3498 rows)

-   `df_tools` contains responses to the tool id and tool confidence questions, where tool_id is a multi-select Q

-   `df_actions` contains responses to the action id and action confidence questions, where df_action is a multi-select Q

-   `df_graphs_full` is a stimulus (ie. graph image) level dataframe where each row contains columns for a participant and all the Qs for that graph

-   `df_graphs` is df_graphs_full w/o the free response and multiselect Qs

## Sanity Check

```{r SANITY-CHECK}

########## CHECK BLOCK COUNTS 
#############################################################################
# title = "Participants by Condition and Data Collection Modality"
# cols = c("Control Condition","Impasse Condition","Total for Period")

title = "Number of abandoned/screened/rejected attempts "
cols = c("Block","Study1","Study2","Study3","Study0","sum")
cont <- table(df_exclude$Assigned.Block, df_exclude$Study)
cont %>% addmargins() %>% kbl(caption = title, col.names = cols) %>%  kable_classic()

title = "Number of abandoned/screened/rejected attempts by type"
cols = c("Rejection Type", "Sum")
cont <- table(df_exclude$EndState)
cont %>% addmargins() %>% kbl(caption = title, col.names = cols) %>%  kable_classic()

title = "Number of successful surveys"
cols = c("Assigned.Block", "Sum")
cont <- table(df_data$Assigned.Block)
cont %>% addmargins() %>% kbl(caption = title, col.names = cols) %>%  kable_classic()

title = "Number of successful surveys by distribution and block"
cols = c("Sampling Platform", "Block-1","Block-2","Block-3","Block-4","Block-5","Block-6", "Sum")
cont <- table(df_data$Distribution, df_data$Assigned.Block)
cont %>% addmargins() %>% kbl(caption = title, col.names = cols) %>%  kable_classic()

title = "Number of successful surveys by sample and block"
cols = c("Sample", "Block-1","Block-2","Block-3","Block-4","Block-5","Block-6", "Sum")
cont <- table(df_data$Sample, df_data$Assigned.Block)
cont %>% addmargins() %>% kbl(caption = title, col.names = cols) %>%  kable_classic()

title = "Number of successful surveys by study and block"
cols = c("Study", "Block-1","Block-2","Block-3","Block-4","Block-5","Block-6", "Sum")
cont <- table(df_data$Study, df_data$Assigned.Block)
cont %>% addmargins() %>% kbl(caption = title, col.names = cols) %>%  kable_classic()

```

## Create MASTER data file

`df_data` represents **response data** for all valid participants (wide format) — answers are not stored as factors; not useful for numeric analysis but just as a record of participant-level responses

```{r CREATE-MASTER}

#WRITE  A CSV FILE WITH RESPONSE DATA FOR VALID PARTICIPANTS (NOT EXCLUDED) in wide format
write.csv(df_data, file = "data/output/Study_1_2/df_data.csv", na="")
saveRDS(df_data, file = "data/output/Study_1_2/df_data.rds")

```

## Create DEMOGRAPHICS file

`df_participants` represents **demographic data** for all valid participants (WIDE)

```{r CREATE-DEMOGRAPHICS}

##### CREATE PARTICIPANT LEVEL DEMOGRAPHIC DATAFRAME 
df_participants <- df_data %>% 
  select(
    PID:Assigned.Block, 
    EndState, Sample, Study, 
    duration.sec, duration.min, 
    contains("D_"), 
    contains("SCREEN_"),
    PLATFORM
  )

#WRITE  A CSV FILE WITH DEMOGRAPHIC DATA ALL VALID PARTICIPANTS (NOT EXCLUDED) in wide format
write.csv(df_participants, file = "data/output/Study_1_2/df_participants.csv", na="")
saveRDS(df_participants, file = "data/output/Study_1_2/df_participants.rds")

```

## Create QDA WIDE

Two files are created for the purpose of qualitative data analysis: - `df_qda_wide`: PARTICIPANT level: includes all valid responses in wide format, to be used for free response data analysis (wide) - `df_qda_long`: GRAPH level: represents response and demographic data for all valid participants (used to generate a spreadsheet for analysis for free response data) *at the trial (i.e. stimulus) level* (i.e. long).

```{r CREATE-QDA-WIDE}

## save participant-level data file for qda validation 
## this does NOT contain excluded participants 
df_qda_wide <- df_data
write.csv(df_data, file = "data/output/Study_1_2/df_qda_wide.csv", na="")

```

## Create QDA LONG

```{r CREATE-QDA-LONG}

#2 CREATE TRIAL LEVEL DFS FOR QDA ###############################################
#### CHART LEVEL DATA FRAME (LONG) FOR QDA (incl demographics) ##################
#### INCLUDES PILOT DATA FROM DATACOLLAR BLUECOLLAR PROLIFIC RECRUITMENT ########
# 1 ROW / participant X GRAPH including demographics 
# UNRAVEL TO QUESTIONS
df_qda_long <- df_data %>% 
  # select(
  # ID.Qualtrics:ID.Study, PLATFORM,
  # contains("_Q_"), contains("loop")
# ) %>% 
  pivot_longer( #PIVOT ON stimulus
  cols = contains("_Q_"),
  names_to = c("stimulus","dummy","BLOCK","QUESTION"),
  values_to = c("value"),
  names_sep = "_"
) %>% select(-dummy) %>% 
  unite(
   BLOCK:stimulus, col="STIMULUS", sep="-", remove=FALSE
) %>% 
  mutate(
    BLOCK = factor(BLOCK),
    STIMULUS = factor(STIMULUS),
    QUESTION = str_replace_all(QUESTION,"-","_"),
    QUESTION = factor(QUESTION),
    STIMULUS_CATEGORY = str_remove(STIMULUS,"B.-"),
    STIMULUS_CATEGORY = factor(STIMULUS_CATEGORY,
                  levels=c("0","4","3","2","1"),
                  labels= c("F","D","C","B","A"))
) %>% 
  select(-stimulus) %>% 
# RE-RAVEL UP TO STIMULI
  filter(!is.na(value)) %>% 
   pivot_wider(
    names_from = QUESTION,
    values_from = value 
  ) %>%  
  tidyr::unnest() %>%  # handle r coerces values to lists
  mutate(
    across(contains("MAKER_ID") | contains("MAKER_GENDER") | contains("MAKER_AGE"), factor),
    across(contains("_CONF") | contains("_LATENCY"), as.numeric),
    across(MAKER_DESIGN:MAKER_TRUST, as.numeric),
    across(CHART_LIKE:CHART_TRUST, as.numeric),
    ENCOUNTER = factor(ENCOUNTER),
    # loop_number = as.numeric(loop_number),
    # loop_number = ifelse(is.na(loop_number), 0, loop_number),
    MAKER_LATENCY = round(MAKER_LATENCY/60,2), #CHANGE TO MINS
    CHART_LATENCY = round(CHART_LATENCY/60,2) #CHANGE TO MINS
  )
#WRITE A CSV FILE AS THE BASIS FOR THE QUALITATIVE DATA ANALYSIS
write.csv(df_qda_long, file = "data/output/Study_1_2/df_qda_long.csv", na="")

```

## Create QUESTION level (long)

`df_questions` is at the graph-question level, but includes *all* questions, both categorical, free response, and numeric.

```{r QUESTION-LONG, warning=FALSE}

#### QUESTION LEVEL DATA FRAME (LONG) ##########################
# unravel ALL the way down to questions 
# 1 row per participant-graph-question
df_questions <- df_data %>% 
  select(
  PID, duration.min,  Assigned.Block,
  Sample, Study, Distribution, PLATFORM, 
  D_gender_collapsed:D_politicsFiscal, 
  contains("_Q_"), contains("loop"), 
) %>% 
  pivot_longer( #PIVOT ON stimulus
  cols = contains("_Q_"),
  names_to = c("stimulus","dummy","BLOCK","QUESTION"),
  values_to = c("value"),
  names_sep = "_"
) %>% select(-dummy) %>% 
  unite(
   BLOCK:stimulus, col="STIMULUS", sep="-", remove=FALSE
) %>% 
  mutate(
    BLOCK = factor(BLOCK),
    STIMULUS = factor(STIMULUS),
    QUESTION = str_replace_all(QUESTION,"-","_"),
    QUESTION = factor(QUESTION, 
                      levels = c(
                        "ENCOUNTER",    
                        "MAKER_ID",        
                        "MAKER_DETAIL", 
                        "MAKER_CONF",
                        "MAKER_AGE", 
                        "AGE_CONF",
                        "MAKER_GENDER",
                        "GENDER_CONF",
                        "MAKER_DESIGN",  
                        "MAKER_DATA",    
                        "MAKER_POLITIC",
                        "MAKER_ARGUE",  
                        "MAKER_SELF",   
                        "MAKER_ALIGN",    
                        "MAKER_TRUST", 
                        "MAKER_EXPLAIN",
                        "MAKER_LATENCY", 
                        "TOOL_ID",     
                        "TOOL_CONF",     
                        "TOOL_DETAIL",   
                        "CHART_LIKE",    
                        "CHART_BEAUTY",  
                        "CHART_INTENT", 
                        "CHART_TRUST",   
                        "CHART_TYPE",    
                        "CHART_ACTION",  
                        "CHART_EXPLAIN", 
                        "CHART_LATENCY", 
                        "loop_number" )),
    STIMULUS_CATEGORY = str_remove(STIMULUS,"B.-"),
    STIMULUS_CATEGORY = factor(STIMULUS_CATEGORY,
                  levels=c("0","4","3","2","1"),
                  labels= c("F","D","C","B","A")),
    tempblock = paste("B",as.character(Assigned.Block), sep="")
) %>% 
  select(-stimulus) %>% #drop unformatted stimulus column
  filter(QUESTION !="loop_number") %>%  #drop loop number rows, we don't use this question
  ## !! FILTER OUT BLANK ROWS (i.e otherwise there is a row for each stimulus in the blocks the person was not assigned to)
  filter(as.character(BLOCK)==as.character(tempblock) | BLOCK=="B0") %>% 
  select(-tempblock) #drop tempblock column



## WRITE DATAFRAME
write.csv(df_questions, file = "data/output/Study_1_2/df_questions_long.csv", na="")
saveRDS(df_questions, file = "data/output/Study_1_2/df_questions_long.rds")
```

## Create SD questions (long)

`df_sd_questions_long` includes only the Semantic Differential questions (still in long format); 1 row per participant-graph-question

```{r SD-LONG}

### SD QUESTIONS LEVEL DATA FRAME (LONG) #####################
## 1 row per participant X sd question X block
## 318p X 11q X 5s = 17,490
df_sd_questions_long <- df_questions %>% 
  filter(QUESTION %in% ref_sd_questions) %>% 
  mutate(
    QUESTION = droplevels(QUESTION),
    value = as.numeric(value)
  )
## WRITE DATAFRAME
write.csv(df_sd_questions_long, file = "data/output/Study_1_2/df_sd_questions_long.csv", na="")
saveRDS(df_sd_questions_long, file = "data/output/Study_1_2/df_sd_questions_long.rds")

```

## Create SD questions (wide)

`df_sd_questions_wide` includes only the Semantic Differential questions (in wide format); 1 row per participant-question; but with a column for each stimulus

```{r SD-WIDE}

#### SD QUESTION LEVEL DATA FRAME (wide-stim) ##########################
# ravel up one level from questions
# 1 row per participant-SDquestion with all blocks as cols for SD qs
# 318p X 11q = 3498 rows
df_sd_questions_wide <- df_questions %>%
  select(-BLOCK,-STIMULUS_CATEGORY) %>% #drop block in order to work at stimulus level
  filter(QUESTION %in% ref_sd_questions) %>% 
  pivot_wider(
    names_from = STIMULUS,
    values_from = value 
  ) %>%  
  tidyr::unnest() %>%  # handle r coerces values to lists
  mutate(
    across(contains("-") , as.numeric),
    QUESTION = droplevels(QUESTION)
  )
## WRITE DATAFRAME
write.csv(df_sd_questions_wide, file = "data/output/Study_1_2/df_sd_questions_wide.csv", na="")
saveRDS(df_sd_questions_wide, file = "data/output/Study_1_2/df_sd_questions_wide.rds")
```

## Create TOOLS

`df_tools` contains 1 row per PARTICIPANT-GRAPH-TOOL (tool is a multi-select question)

```{r TOOLS}

## DF_TOOLS 
# 1 ROW per PARTICIPANT X STIMULUS X tool_id selection 
# tool_id is a multiselect field 
df_tools <- df_questions %>% 
  # select(ID.Prolific, QUESTION, value) %>%
  filter(QUESTION  %in% c("TOOL_ID", "TOOL_CONF", 
                          "MAKER_DESIGN"  ,"MAKER_DATA" ,"MAKER_POLITIC" ,"MAKER_ARGUE"  ,"MAKER_SELF"    ,"MAKER_ALIGN"  ,"MAKER_TRUST",
                          "CHART_LIKE","CHART_BEAUTY"  ,"CHART_INTENT", "CHART_TRUST",
                          "MAKER_ID", "MAKER_AGE","MAKER_GENDER"
                          )) %>% 
  mutate(QUESTION = fct_drop(QUESTION)) %>% 
  pivot_wider(
    names_from = QUESTION,
    values_from = value 
  ) %>% 
  separate_longer_delim(
    cols = TOOL_ID,
    delim = ","
  ) %>% 
  mutate(TOOL_ID = factor(TOOL_ID, 
                             levels = order_tools),
         PID = droplevels(PID),
         TOOL_CONF = as.numeric(TOOL_CONF), 
         MAKER_DESIGN  = as.numeric(MAKER_DESIGN  ),
         MAKER_DATA= as.numeric(MAKER_DATA),
         MAKER_POLITIC = as.numeric(MAKER_POLITIC ),
         MAKER_ARGUE  = as.numeric(MAKER_ARGUE  ),
         MAKER_SELF= as.numeric(MAKER_SELF),
         MAKER_ALIGN  = as.numeric(MAKER_ALIGN  ),
         MAKER_TRUST= as.numeric(MAKER_TRUST),
         CHART_LIKE   = as.numeric(CHART_LIKE   ),
         CHART_BEAUTY  = as.numeric(CHART_BEAUTY  ),
         CHART_INTENT = as.numeric(CHART_INTENT ),
         CHART_TRUST = as.numeric(CHART_TRUST ),
         MAKER_ID = factor(MAKER_ID, levels = order_maker),
         MAKER_AGE = factor(MAKER_AGE, levels = order_age),
         MAKER_GENDER = factor(MAKER_GENDER, levels = order_gender),
        ) 
## WRITE DATAFRAME
write.csv(df_tools, file = "data/output/Study_1_2/df_tools.csv", na="")
saveRDS(df_tools, file = "data/output/Study_1_2/df_tools.rds")
```

## Create ACTIONS

`df_actions` contains 1 row per PARTICIPANT-GRAPH-ACTION (tool is a multi-select question)

```{r ACTIONS}
## DF_ACTIONS 
# 1 ROW per PARTICIPANT X STIMULUS X chart_action selection 
# chart_action is a multiselect field 
df_actions <- df_questions %>% 
  # filter(QUESTION == "CHART_ACTION") %>%
  filter(QUESTION  %in% c("CHART_ACTION",
                          "MAKER_DESIGN"  ,"MAKER_DATA" ,"MAKER_POLITIC" ,"MAKER_ARGUE"  ,"MAKER_SELF"    ,"MAKER_ALIGN"  ,"MAKER_TRUST",
                          "CHART_LIKE","CHART_BEAUTY"  ,"CHART_INTENT", "CHART_TRUST","MAKER_ID", "MAKER_AGE","MAKER_GENDER" )) %>%
  mutate(QUESTION = fct_drop(QUESTION)) %>% 
  pivot_wider(
    names_from = QUESTION,
    values_from = value 
  ) %>% 
  separate_longer_delim(
    # cols = value,
    cols = CHART_ACTION,
    delim = ","
  ) %>% 
  # mutate(CHART_ACTION = factor(value, 
  mutate(CHART_ACTION = factor(CHART_ACTION, 
                                  levels =  c("NOTHING — just keep scrolling",
                                              "unfollow / block the source",
                                              "post a comment",
                                              "share / repost",
                                              "share / repost WITH comment",
                                              "look up more information about the topic or source"),
                                  labels =  c("nothing",
                                              "unfollow/block",
                                              "comment",  
                                              "share",
                                              "share w/ comment",
                                              "seek information")),
         CHART_ACTION = factor(CHART_ACTION, levels = order_actions),
         
         ## CONDENSE FOR 2 CATEGORIES
         CHART_ACTION2 = fct_recode(CHART_ACTION ,
         nothing  = "nothing",
         something = "unfollow/block",
         something = "comment",  
         something = "share",
         something = "share w/ comment",
         something = "seek information"),
         CHART_ACTION2 = factor(CHART_ACTION2, levels = order_actions2),

         
         
         ## CONDENSE FOR 3 CATEGORIES
         CHART_ACTION3 = fct_recode(CHART_ACTION ,
         nothing  = "nothing",
         solo_engagement = "unfollow/block",
         social_engagement = "comment",  
         social_engagement = "share",
         social_engagement = "share w/ comment",
         solo_engagement = "seek information"),
         # CHART_ACTION3 = factor(CHART_ACTION3, levels = order_actions3),
         
         
         ## CONDENSE FOR 4 CATEGORIES
         CHART_ACTION4 = fct_recode(CHART_ACTION ,
         nothing  = "nothing",
         unfollow_block = "unfollow/block",
         comment_share = "comment",  
         comment_share = "share",
         comment_share = "share w/ comment",
         seek_info = "seek information"),
         # CHART_ACTION4 = factor(CHART_ACTION4, levels = order_actions4),
      
         
         MAKER_DESIGN  = as.numeric(MAKER_DESIGN  ),
         MAKER_DATA= as.numeric(MAKER_DATA),
         MAKER_POLITIC = as.numeric(MAKER_POLITIC ),
         MAKER_ARGUE  = as.numeric(MAKER_ARGUE  ),
         MAKER_SELF= as.numeric(MAKER_SELF),
         MAKER_ALIGN  = as.numeric(MAKER_ALIGN  ),
         MAKER_TRUST= as.numeric(MAKER_TRUST),
         CHART_LIKE   = as.numeric(CHART_LIKE   ),
         CHART_BEAUTY  = as.numeric(CHART_BEAUTY  ),
         CHART_INTENT = as.numeric(CHART_INTENT ),
         CHART_TRUST = as.numeric(CHART_TRUST ),
         MAKER_ID = factor(MAKER_ID, levels = order_maker),
         MAKER_AGE = factor(MAKER_AGE, levels = order_age),
         MAKER_GENDER = factor(MAKER_GENDER, levels = order_gender),
         PID = droplevels(PID)) 
  # select(-value,-QUESTION)
## WRITE DATAFRAME
write.csv(df_actions, file = "data/output/Study_1_2/df_actions.csv", na="")
saveRDS(df_actions, file = "data/output/Study_1_2/df_actions.rds")
```

## Create GRAPH/TRIAL level (all questions)

`df_graphs_full` contains 1 row / participant-graph (all qs as columns). The \_full indicates the file contains free response questions (useful for hover text on interactive visualizations)

```{r GRAPH-LEVEL-FULL}
#### DFS FOR TRIAL LEVEL ANALYSIS   ######################################################
#### CHART LEVEL DATA FRAME (LONG) #####################################################
# roll partway back up from questions 
# 1 row per participant X graph 
# unnest https://stackoverflow.com/questions/58035452/pivot-wider-outputs-s3-vctrs-list-of-objects
df_graphs_full <- df_questions %>% 
  pivot_wider(
    names_from = QUESTION,
    values_from = value 
  ) %>%  
  tidyr::unnest() %>%  # handle r coerces values to lists
  mutate(
    across(contains("MAKER_ID") | contains("MAKER_GENDER") | contains("MAKER_AGE"), factor),
    across(contains("_CONF") | contains("_LATENCY"), as.numeric),
    across(MAKER_DESIGN:MAKER_TRUST, as.numeric),
    across(CHART_LIKE:CHART_TRUST, as.numeric),
    ENCOUNTER = factor(ENCOUNTER),
    # loop_number = as.numeric(loop_number),
    # loop_number = ifelse(is.na(loop_number), 0, loop_number),
    MAKER_LATENCY = round(MAKER_LATENCY/60,2), #CHANGE TO MINS
    CHART_LATENCY = round(CHART_LATENCY/60,2), #CHANGE TO MINS
    MAKER_ID = factor( MAKER_ID,levels = c("business", "political", "education","news","organization","individual"))
  )
## WRITE DATAFRAME
write.csv(df_graphs_full, file = "data/output/Study_1_2/df_graphs_full.csv", na="")
saveRDS(df_graphs_full, file = "data/output/Study_1_2/df_graphs_full.rds")

```

## Create GRAPH/TRIAL level (SD and categorical)

`df_graphs_full` contains 1 row / participant-graph (all qs as columns), only for the SD and categorical (not multi-select or free response questions).

```{r GRAPH-LEVEL-MAIN}

#### DFS FOR TRIAL LEVEL ANALYSIS w/o free response ######################################################
#### CHART LEVEL DATA FRAME (LONG) ##########################
## SUBSET OF COLUMNS EXCLUDING THE FREE RESPONSES and MULTISELECT
df_graphs <- df_graphs_full %>% 
  select( !where(is.character)) %>% 
  mutate(
    ##re order factor levels
   MAKER_ID = factor(MAKER_ID, levels = order_maker),
   MAKER_AGE = factor(MAKER_AGE, levels = order_age),
   MAKER_GENDER = factor(MAKER_GENDER, levels = order_gender),
   ENCOUNTER = factor(ENCOUNTER, levels = order_encounter),
  ) 
## WRITE DATAFRAME
write.csv(df_graphs, file = "data/output/Study_1_2/df_graphs.csv", na="")
saveRDS(df_graphs, file = "data/output/Study_1_2/df_graphs.rds")

```

## Z-SCORED SD

Create duplicate versions of datasets used to analyze the semantic differential scores, with standardized (i.e. z-scored). Created as separate files rather than adding cols to extant dfs b/c it makes the looping graphing code more readable

```{r, standardizing-zscores}

########## Z-SCORE
## ON STANDARDIZATION
## https://cran.r-project.org/web/packages/datawizard/vignettes/standardize_data.html

## DF_GRAPHS_Z
## VARIABLE WISE STANDARDIZATION
df_graphs_z <- df_graphs %>% 
  mutate(
    MAKER_DESIGN_z  = datawizard::standardize(MAKER_DESIGN ),
    MAKER_DATA_z = datawizard::standardize(MAKER_DATA),
    MAKER_POLITIC_z = datawizard::standardize(MAKER_POLITIC),
    MAKER_ARGUE_z = datawizard::standardize(MAKER_ARGUE),
    MAKER_SELF_z = datawizard::standardize(MAKER_SELF),
    MAKER_ALIGN_z = datawizard::standardize(MAKER_ALIGN),
    MAKER_TRUST_z = datawizard::standardize(MAKER_TRUST),
    CHART_LIKE_z = datawizard::standardize(CHART_LIKE),
    CHART_BEAUTY_z = datawizard::standardize(CHART_BEAUTY),
    CHART_INTENT_z = datawizard::standardize(CHART_INTENT),
    CHART_TRUST_z = datawizard::standardize(CHART_TRUST),
    MAKER_CONF_z = datawizard::standardize(MAKER_CONF),
    AGE_CONF_z = datawizard::standardize(AGE_CONF),
    GENDER_CONF_z = datawizard::standardize(GENDER_CONF),
    TOOL_CONF_z = datawizard::standardize(TOOL_CONF)
  )
write.csv(df_graphs_z, file = "data/output/Study_1_2/df_graphs_z.csv", na="")
saveRDS(df_graphs_z, file = "data/output/Study_1_2/df_graphs_z.rds")

## DF_SD_QUESTIONS_LONG_Z
## LONG DATAFRAME FOR RAINCLOUDS
df_sd_questions_long_z <- df_graphs_z %>% 
  ##first transform dw-transformer to numeric
  mutate(
    MAKER_DESIGN_z = as.numeric(MAKER_DESIGN_z),
    MAKER_DATA_z = as.numeric(MAKER_DATA_z),
    MAKER_POLITIC_z = as.numeric(MAKER_POLITIC_z),
    MAKER_ARGUE_z = as.numeric(MAKER_ARGUE_z),
    MAKER_SELF_z = as.numeric(MAKER_SELF_z),
    MAKER_ALIGN_z = as.numeric(MAKER_ALIGN_z),
    MAKER_TRUST_z = as.numeric(MAKER_TRUST_z),
    CHART_LIKE_z = as.numeric(CHART_LIKE_z),
    CHART_BEAUTY_z = as.numeric(CHART_BEAUTY_z),
    CHART_INTENT_z = as.numeric(CHART_INTENT_z),
    CHART_TRUST_z = as.numeric(CHART_TRUST_z)) %>% 
  #drop non z-scored columns AND confidence cols
  select(- (ENCOUNTER:CHART_LATENCY), 
         - contains("_CONF"))%>% 
  ## pivot longer the SD questions
  pivot_longer(
    cols = ref_sd_questions_z,
    names_to = "QUESTION"
  ) %>% mutate(
    QUESTION = factor(QUESTION, levels=ref_sd_questions_z)
  )
write.csv(df_sd_questions_long_z, file = "data/output/Study_1_2/df_sd_questions_long_z.csv", na="")
saveRDS(df_sd_questions_long_z, file = "data/output/Study_1_2/df_sd_questions_long_z.rds")    



```

## ABSOLUTE_VALUE SD

Create duplicate versions of datasets used to analyze the semantic differential scores with absolute values (i.e. folding SD scale in half, so 0 = center of the slider, and 50 = the left and right extremes) Created as separate files rather than adding cols to extant dfs b/c it makes the looping graphing code more readable

```{r, standardizing-absvalue}

########## ABS VALUE
## COLLAPSE SD SCALES AT 0 TO SHOW MIDDLE VS EXTREMES

## DF_GRAPHS_ABS
## VARIABLE WISE STANDARDIZATION
df_graphs_abs <- df_graphs %>% 
  mutate(
    MAKER_DESIGN  =  abs(MAKER_DESIGN - 50),
    MAKER_DATA =  abs(MAKER_DATA - 50),
    MAKER_POLITIC =  abs(MAKER_POLITIC - 50),
    MAKER_ARGUE =  abs(MAKER_ARGUE - 50),
    MAKER_SELF =  abs(MAKER_SELF - 50),
    MAKER_ALIGN =  abs(MAKER_ALIGN - 50),
    MAKER_TRUST =  abs(MAKER_TRUST - 50),
    CHART_LIKE =  abs(CHART_LIKE - 50),
    CHART_BEAUTY =  abs(CHART_BEAUTY - 50),
    CHART_INTENT =  abs(CHART_INTENT - 50),
    CHART_TRUST =  abs(CHART_TRUST - 50),
    MAKER_CONF =  abs(MAKER_CONF - 50),
    AGE_CONF =  abs(AGE_CONF - 50),
    GENDER_CONF =  abs(GENDER_CONF - 50),
    TOOL_CONF =  abs(TOOL_CONF - 50)
    
  )
write.csv(df_graphs_abs, file = "data/output/Study_1_2/df_graphs_abs.csv", na="")
saveRDS(df_graphs_abs, file = "data/output/Study_1_2/df_graphs_abs.rds")

## DF_SD_QUESTIONS_LONG_ABS
## LONG DATAFRAME FOR RAINCLOUDS
df_sd_questions_long_abs <- df_graphs_abs %>% 
  ## pivot longer the SD questions
  pivot_longer(
    cols = ref_sd_questions,
    names_to = "QUESTION"
  ) %>% mutate(
    QUESTION = factor(QUESTION, levels=ref_sd_questions)
  )
write.csv(df_sd_questions_long_abs, file = "data/output/Study_1_2/df_sd_questions_long_abs.csv", na="")
saveRDS(df_sd_questions_long_abs, file = "data/output/Study_1_2/df_sd_questions_long_abs.rds")    



```

# PROFILING

## PARTICIPANT-LEVEL

```{r summarytools-profile-participants, results='asis'}


dfSummary(df_participants , #%>% select(PID, duration.min, Distribution, Assigned.Block, PLATFORM, 
                      #       contains("D_"), Prolific.Name:Study, contains("SCREEN_")), 
          headings = TRUE,
          plain.ascii  = FALSE,
          style        = 'grid',
          graph.magnif = 0.85,
          varnumbers = FALSE,
          valid.col    = FALSE,
          tmp.img.dir  = "/tmp")


```

## TRIAL-LEVEL

```{r summarytools-profile-graphs, results='asis'}


dfSummary(df_graphs,  # %>% select(PID, duration.min, Distribution, Assigned.Block, PLATFORM,
                      #       contains("D_"), Prolific.Name:Study, contains("SCREEN_")), 
          headings = TRUE,
          plain.ascii  = FALSE,
          style        = 'grid',
          graph.magnif = 0.85,
          varnumbers = FALSE,
          valid.col    = FALSE,
          tmp.img.dir  = "/tmp")


```

# STASH

```{r}
# EXPLORING HOW CORRELATION IS AFFECTED BY STANDARDIZING BY VARIABLE, PARTICIPANT OR BOTH
# AS WELL AS THE EFFECT OF SUBJECT RANDOM EFFECT IN CORRPLOT
# 
# ########################################################################
# ### CORRELATION OF RAW SCORES (NO RANDOM ID)
# df <- df_graphs %>% select(
#           MAKER_DESIGN, MAKER_DATA, MAKER_POLITIC, MAKER_ARGUE,
#           MAKER_SELF, MAKER_ALIGN, MAKER_TRUST, 
#           CHART_LIKE, CHART_BEAUTY, CHART_INTENT, CHART_TRUST)
# 
# print("PARTIAL CORRELATION")
# #CALCULATE partial correlations 
# c <- df %>% 
#   correlation(partial=TRUE)
# 
# #PRINT A SUMMARY
# (s1 <- c %>% summary(redundant = FALSE ))
# #GET THE MATRIX
# m <- as.matrix(c)
# #VIS WITH CORRPLOT
# 
# 
# ## SHADED NUMBERS
# corrplot(m,  order = 'original', method = "number", 
#          diag = FALSE, type = "lower",
#          insig='blank',
#          # insig = 'label_sig', sig.level = c(0.001, 0.01, 0.05),
#          addCoef.col = '#595D60',
#          tl.pos = "ld", tl.col = "#595D60",
#          title = "CORRELATION OF RAW SCORES (NO RANDOM ID)") 
# 
# 
# ########################################################################
# ### CORRELATION OF RAW SCORES W/ RANDOM PID
# df <- df_graphs %>% select(
#           MAKER_DESIGN, MAKER_DATA, MAKER_POLITIC, MAKER_ARGUE,
#           MAKER_SELF, MAKER_ALIGN, MAKER_TRUST, 
#           CHART_LIKE, CHART_BEAUTY, CHART_INTENT, CHART_TRUST, PID)
# 
# print("PARTIAL CORRELATION")
# #CALCULATE partial correlations with PID as random effect
# c <- df %>% 
#   correlation(partial=TRUE, multilevel = TRUE)
# 
# #PRINT A SUMMARY
# (s2 <- c %>% summary(redundant = FALSE ))
# #GET THE MATRIX
# m <- as.matrix(c)
# #VIS WITH CORRPLOT
# 
# 
# ## SHADED NUMBERS
# corrplot(m,  order = 'original', method = "number", 
#          diag = FALSE, type = "lower",
#          insig='blank',
#          # insig = 'label_sig', sig.level = c(0.001, 0.01, 0.05),
#          addCoef.col = '#595D60',
#          tl.pos = "ld", tl.col = "#595D60",
#          title = "CORRELATION OF RAW SCORES W/ RANDOM PID") 
# 
# 
# 
# ########################################################################
# ### CORRELATION OF VARIABLE WISE Z-SCORED SCORES NO RANDOM ID
# df <- df_graphs_z %>% select(
#           MAKER_DESIGN, MAKER_DATA, MAKER_POLITIC, MAKER_ARGUE,
#           MAKER_SELF, MAKER_ALIGN, MAKER_TRUST, 
#           CHART_LIKE, CHART_BEAUTY, CHART_INTENT, CHART_TRUST)
# 
# print("PARTIAL CORRELATION WITH PID AS RANDOM EFFECT ON Z-SCORED SDS")
# #CALCULATE partial correlations with PID as random effect
# c <- df %>% 
#   correlation(partial=TRUE)
# 
# #PRINT A SUMMARY
# (s3 <- c %>% summary(redundant = FALSE ))
# #GET THE MATRIX
# m <- as.matrix(c)
# #VIS WITH CORRPLOT
# 
# 
# ## SHADED NUMBERS
# corrplot(m,  order = 'original', method = "number", 
#          diag = FALSE, type = "lower",
#          insig='blank',
#          # insig = 'label_sig', sig.level = c(0.001, 0.01, 0.05),
#          addCoef.col = '#595D60',
#          tl.pos = "ld", tl.col = "#595D60",
#          title = "CORRELATION OF VARIABLE WISE Z-SCORED SCORES NO RANDOM ID") 
# 
# ########################################################################
# ### CORRELATION OF VARIABLE WISE Z-SCORED SCORES W/ RANDOM PID
# df <- df_graphs_z %>% select(
#           MAKER_DESIGN, MAKER_DATA, MAKER_POLITIC, MAKER_ARGUE,
#           MAKER_SELF, MAKER_ALIGN, MAKER_TRUST, 
#           CHART_LIKE, CHART_BEAUTY, CHART_INTENT, CHART_TRUST, 
#           PID)
# 
# print("PARTIAL CORRELATION WITH PID AS RANDOM EFFECT ON Z-SCORED SDS")
# #CALCULATE partial correlations with PID as random effect
# c <- df %>% 
#   correlation(partial=TRUE,
#               multilevel = TRUE)
# 
# #PRINT A SUMMARY
# (s4 <- c %>% summary(redundant = FALSE ))
# #GET THE MATRIX
# m <- as.matrix(c)
# #VIS WITH CORRPLOT
# 
# 
# ## SHADED NUMBERS
# corrplot(m,  order = 'original', method = "number", 
#          diag = FALSE, type = "lower",
#          insig='blank',
#          # insig = 'label_sig', sig.level = c(0.001, 0.01, 0.05),
#          addCoef.col = '#595D60',
#          tl.pos = "ld", tl.col = "#595D60",
#          title = "CORRELATION OF VARIABLE WISE Z-SCORED SCORES W/ RANDOM PID") 
# 
# 
# ########################################################################
# ### CORRELATION OF (BOTH) WISE Z-SCORED SCORES W/0 RANDOM PID
# df <- df_graphs_zb %>% select(
#           MAKER_DESIGN, MAKER_DATA, MAKER_POLITIC, MAKER_ARGUE,
#           MAKER_SELF, MAKER_ALIGN, MAKER_TRUST, 
#           CHART_LIKE, CHART_BEAUTY, CHART_INTENT, CHART_TRUST)
# 
# print("PARTIAL CORRELATION WITH PID AS RANDOM EFFECT ON Z-SCORED SDS")
# #CALCULATE partial correlations with PID as random effect
# c <- df %>% 
#   correlation(partial=TRUE)
#               
# 
# #PRINT A SUMMARY
# (s5 <- c %>% summary(redundant = FALSE ))
# #GET THE MATRIX
# m <- as.matrix(c)
# #VIS WITH CORRPLOT
# 
# 
# ## SHADED NUMBERS
# corrplot(m,  order = 'original', method = "number", 
#          diag = FALSE, type = "lower",
#          insig='blank',
#          # insig = 'label_sig', sig.level = c(0.001, 0.01, 0.05),
#          addCoef.col = '#595D60',
#          tl.pos = "ld", tl.col = "#595D60",
#          title = "CORRELATION OF (BOTH) WISE Z-SCORED SCORES W/0 RANDOM PID") 
# 
# 
# 
# ########################################################################
# ### CORRELATION OF (BOTH) WISE Z-SCORED SCORES RANDOM PID
# df <- df_graphs_zb %>% select(
#           MAKER_DESIGN, MAKER_DATA, MAKER_POLITIC, MAKER_ARGUE,
#           MAKER_SELF, MAKER_ALIGN, MAKER_TRUST, 
#           CHART_LIKE, CHART_BEAUTY, CHART_INTENT, CHART_TRUST, 
#           PID)
# 
# print("PARTIAL CORRELATION ON Z-SCORED SDS")
# #CALCULATE partial correlations (no random effect b/c scores are participant-wise z-scored)
# c <- df %>% 
#   correlation(partial=TRUE, multilevel = TRUE)
# 
# #PRINT A SUMMARY
# (s6s1 <- c %>% summary(redundant = FALSE ))
# #GET THE MATRIX
# m <- as.matrix(c)
# #VIS WITH CORRPLOT
# 
# 
# ## SHADED NUMBERS
# corrplot(m,  order = 'original', method = "number", 
#          diag = FALSE, type = "lower",
#          insig='blank',
#          # insig = 'label_sig', sig.level = c(0.001, 0.01, 0.05),
#          addCoef.col = '#595D60',
#          tl.pos = "ld", tl.col = "#595D60",
#          title = "CORRELATION OF (BOTH) WISE Z-SCORED SCORES RANDOM PID") 
# 
# 
# 
# 
# 



```
