==> Studies/S2-surveys/2_surveys/3_VIS25_submission/0.2_SUB1006_combined_wrangling.Rmd <==
---
title: "COMBINED-DATA WRANGLING—SUBMISSION 1006"
author: "ANONYMIZED"
date: "2025-02-24"
output:
  html_document:
    theme: flatly
    code_folding: hide
    fig_caption: yes
    number_sections: yes
    toc: yes
    toc_float:
      collapsed: yes
      smooth_scroll: yes
    toc_depth: 5
  pdf_document:
    toc: yes
    toc_depth: '4'
always_allow_html: yes
font-family: DejaVu Sans
mainfont: DejaVu Sans
---

```{r setup, include=FALSE, message=FALSE}
knitr::opts_chunk$set(echo = TRUE)

#UTILITIES
library(Hmisc) # %nin% operator
library(psych) #describe()
library(tidyverse) #all the things
library(magrittr) #special pipes like %<>%
library(lubridate) #dealing with dates
library(kableExtra) #printing tables

#EXPLORATORY DATA ANALYSIS
library(summarytools) #data quality
library(qacBase) #EDA

#CONFIG
options(readr.show_col_types = FALSE) #don't show coltypes on read_csv
n_blocks = 6 #number of stimulus blocks in study design

```

*This notebook contains code to combine data from Study 0, Study 1, Study 2, and Study 3 of the MAGIC-VIS project to a single single normalized dataset of (valid) participant responses.* 

**NOTE** This file can only be executed _after_ the study specific wrangling files have been run, as those notebooks transform the data from raw qualtrics format to a standardized format. 

**NOTE** Data from Study 3 are only included at the 'pre-test' (graph-text-obscured) stage. Data from the post-test are excluded, as this data was not gathered in the other studies.  

Four types of files are generated: 
(1) Participant-level : 1 row per participant
(2) Graph-level : 1 row per participantXgraph (5 graphs per participant; except Study 3, 5 graphs per participant X 2 times (pre-post))
(3) Question-level : 1 row per participantXgraphXquestion 

## IMPORT Configuration Files

*Import files with reference information for stimuli `ref_stimuli` and qualtrics surveys `ref_surveys`.*

```{r setup-references, message=FALSE, warning = FALSE}

############## IMPORT STIMULI FILE
#contains data for each stimulus used 
ref_stimuli <- read_csv("data/input/REFERENCE/stimuli.csv", col_names = TRUE) %>% 
  mutate(
    BLOCK = as.factor(BLOCK), 
    STIMULUS_CATEGORY = as.factor(CATEGORY),
    ID = as.factor(ID),
    MAKER_ID = as.factor(MAKER_ID)) %>% 
  select(-CATEGORY)



############## STUDY ID FILE
#most blocks were run as separate qualtrics surveys with diffferent recruitments in Prolific
ref_surveys <- read_csv("data/input/REFERENCE/studies.csv", col_names = TRUE) %>%
  mutate(
    ID.Study = as.factor(ID.Study),
    Assigned.Block = as.factor(Assigned.Block),
    Distribution = as.factor(Distribution),
    Prolific.Name = as.factor(Prolific.Name),
    Qualtrics.URL = as.factor(Qualtrics.URL),
    Qualtrics.Survey = as.factor(Qualtrics.Survey),
    Sample = as.factor(Sample),
    Study = as.factor(Study)
)


```

## DEFINE Labels

*Define labels used in survey questions and response options; used to generate labels for graphs and filtering of datasets in analysis files.*

```{r setup-labels}

############## BUILD LABELS
ref_stim_id <- levels(ref_stimuli$ID)

### SD questions on Studies S1, S2
ref_sd_questions <- c("MAKER_DESIGN","MAKER_DATA",
                      "MAKER_POLITIC", "MAKER_ARGUE","MAKER_SELF","MAKER_ALIGN","MAKER_TRUST",
                      "CHART_TRUST", "CHART_INTENT", "CHART_LIKE", "CHART_BEAUTY")
ref_sd_questions_z <- c("MAKER_DESIGN_z","MAKER_DATA_z",
                      "MAKER_POLITIC_z", "MAKER_ARGUE_z","MAKER_SELF_z","MAKER_ALIGN_z","MAKER_TRUST_z",
                      "CHART_TRUST_z", "CHART_INTENT_z", "CHART_LIKE_z", "CHART_BEAUTY_z")
ref_conf_questions <- c("MAKER_CONF","AGE_CONF","GENDER_CONF")
## SD questions on Study3
ref_min_sd_questions <-  c("DESIGN","DATA","POLITICS","ALIGN","INTENT","TRUST","BEAUTY")
ref_min_sd_questions_z <-  c("DESIGN_z","DATA_z","POLITICS_z", "ALIGN_z","INTENT_z","TRUST_z","BEAUTY_z")
ref_min_conf_questions <- c("ID_CONF","AGE_CONF","GENDER_CONF")


### USEFUL CODE FOR TROUBLESHOOTING COMPARING DF STRUCTURE
#dataframe column comparisons 
# x <- janitor::compare_df_cols(df1, df3)

```

## IMPORT Wrangled Data

We want to include in this dataset only responses from participants who saw _message-obscured_ stimuli. Thus we include participants from Studies 1-2, and for Study 3, on the _pre-test_ scores, which reflect answers to questions when participants viewed the message-obscured version of the graph. 
```{r import-wrangled}

## IMPORT STUDIES 1-3 FILES
df_participants_s12 <- readRDS("data/output/Study_1_2/df_participants.rds")
df_graphs_s12 <- readRDS("data/output/Study_1_2/df_graphs_z.rds") #zscore version has both raw and zscores
df_questions_long_s12 <- readRDS("data/output/Study_1_2/df_questions_long.rds") #all qs long
df_sd_questions_long_s12 <- readRDS("data/output/Study_1_2/df_sd_questions_long.rds")
df_sd_questions_long_s12_z <- readRDS("data/output/Study_1_2/df_sd_questions_long_z.rds")


## IMPORT Study 3 FILES
df_participants_s3 <- readRDS("data/output/Study_3/df_participants_S3.rds")
df_graphs_s3 <- readRDS("data/output/Study_3/df_graphs_z_S3.rds") #zscore version has both raw and zscores
df_questions_long_s3 <- readRDS("data/output/Study_3/df_questions_long_S3.rds")
df_sd_questions_long_s3 <- readRDS("data/output/Study_3/df_sd_questions_long_S3.rds")
df_sd_questions_long_s3_z <- readRDS("data/output/Study_3/df_sd_questions_long_z_S3.rds")
```




## JOIN PARTICIPANT level files
```{r join-participants}


## ASSIGN DFS
df1 <- df_participants_s12
df3 <- df_participants_s3

#dataframe column comparisons 
# x <- janitor::compare_df_cols(df1, df3)
#x

## JOIN DFS 
df <- rbind(df1,df3)


## PRINT
print("PARTICIPANTS BY STUDY AND BLOCK")
addmargins(table(df$Study, df$Assigned.Block))

## WRITE FILE
saveRDS(df, file = "data/output/COMBINED/df_participants_ALL.rds")

df_participants_all <- df
## CLEANUP
rm(df, df1, df3)

```


## JOIN GRAPH level Files 

```{r join-graphs}

## WRANGLE STUDIES 1_3, 
df1 <- df_graphs_s12 %>% 
  # Retain only cols needed for modelling
  select(PID, Assigned.Block, Study, Sample, D_politicsSocial, D_politicsFiscal, 
         STIMULUS_CATEGORY, STIMULUS, BLOCK, ENCOUNTER, MAKER_ID, MAKER_AGE, MAKER_GENDER, 
         MAKER_CONF, AGE_CONF, GENDER_CONF,
         MAKER_DESIGN:CHART_TRUST,
         MAKER_CONF_z, AGE_CONF_z,GENDER_CONF_z,
         MAKER_DESIGN_z:CHART_TRUST_z
  ) %>% 
  # Filter out questions dropped from Study 3
  select(-c("MAKER_ARGUE_z","MAKER_SELF_z","CHART_LIKE_z","CHART_TRUST_z"))  %>% 
  select(-c("MAKER_ARGUE","MAKER_SELF","CHART_LIKE","CHART_TRUST","MAKER_LATENCY","TOOL_CONF")) %>%
  # Rename questions to match Study 3 and get rid of magicwizard transform for raw z value
  mutate(
    DESIGN_z = as.numeric(MAKER_DESIGN_z),
    DATA_z = as.numeric(MAKER_DATA_z),
    POLITICS_z = as.numeric(MAKER_POLITIC_z),
    ALIGN_z = as.numeric(MAKER_ALIGN_z),
    TRUST_z = as.numeric(MAKER_TRUST_z),
    BEAUTY_z = as.numeric(CHART_BEAUTY_z),
    INTENT_z = as.numeric(CHART_INTENT_z),
    MAKER_CONF_z = as.numeric(MAKER_CONF_z),
    AGE_CONF_z = as.numeric(AGE_CONF_z),
    GENDER_CONF_z = as.numeric(GENDER_CONF_z)
    ) %>% 
  rename(
    ID = MAKER_ID,
    AGE = MAKER_AGE,
    GENDER = MAKER_GENDER,
    ID_CONF = MAKER_CONF,
    DESIGN = MAKER_DESIGN,
    DATA = MAKER_DATA,
    POLITICS = MAKER_POLITIC,
    ALIGN=MAKER_ALIGN,
    INTENT=CHART_INTENT,
    TRUST=MAKER_TRUST,
    BEAUTY=CHART_BEAUTY,
    ID_CONF_z = MAKER_CONF_z
    
  ) %>% 
  # Drop old columns
  select(-c(MAKER_DESIGN_z:CHART_INTENT_z)) %>%
  mutate(TIME="PRE") %>% ## SET ALL STUDY 1,2 GRAPHS TO PRE B/C the obscured stimulus was presented
  # Drop levels
  droplevels()

## Study 3, RETAIN ONLY COLUMNS NECESSARY FOR MODELLING, and only TIME=PRE
df3_PRE <- df_graphs_s3%>% 
  # Retail only cols needed for modelling
  select(PID, Assigned.Block, Study, Sample, D_politicsSocial, D_politicsFiscal,
        STIMULUS_CATEGORY, STIMULUS, BLOCK,
        # only keep pre-test columns
        contains("PRE")) %>%
  # drop latency
  select(-contains("LATENCY")) %>% 
  # rename categoricals
  rename(
    ENCOUNTER = PRE_ENCOUNTER,
    ID = PRE_ID,
    AGE = PRE_AGE,
    GENDER = PRE_GENDER,
    DESIGN = PRE_DESIGN,
    DATA = PRE_DATA,
    POLITICS = PRE_POLITICS,
    ALIGN = PRE_ALIGN,
    INTENT = PRE_INTENT,
    TRUST = PRE_TRUST,
    BEAUTY = PRE_BEAUTY,
    ID_CONF = PRE_ID_CONF,
    AGE_CONF = PRE_AGE_CONF,
    GENDER_CONF = PRE_GENDER_CONF
  ) %>% 
  #get rid of magicwizard transform for raw z value
  mutate(
    DESIGN_z = as.numeric(PRE_DESIGN_z),
    DATA_z = as.numeric(PRE_DATA_z),
    POLITICS_z = as.numeric(PRE_POLITICS_z),
    ALIGN_z = as.numeric(PRE_ALIGN_z),
    TRUST_z = as.numeric(PRE_TRUST_z),
    BEAUTY_z = as.numeric(PRE_BEAUTY_z),
    INTENT_z = as.numeric(PRE_INTENT_z),
    ID_CONF_z = as.numeric(PRE_ID_CONF_z),
    AGE_CONF_z = as.numeric(PRE_AGE_CONF_z),
    GENDER_CONF_z = as.numeric(PRE_GENDER_CONF_z)
    ) %>% 
  # drop old columns datawizard columns
  select(-c(PRE_DESIGN_z: PRE_INTENT_z)) %>%
  select(-c(PRE_ID_CONF_z, PRE_AGE_CONF_z, PRE_GENDER_CONF_z)) %>% 
  mutate(TIME = "PRE") %>% 
  droplevels()


# Study 3, RETAIN ONLY COLUMNS NECESSARY FOR MODELLING, and only TIME=post
df3_POST <- df_graphs_s3%>% 
  # Retail only cols needed for modelling
  select(PID, Assigned.Block, Study, Sample, D_politicsSocial, D_politicsFiscal,
        STIMULUS_CATEGORY, STIMULUS, BLOCK,
        # only keep pre-test columns
        contains("POST")) %>%
  # drop latency
  select(-contains("LATENCY")) %>% 
  # rename categoricals
  rename(
    ENCOUNTER = POST_ENCOUNTER,
    ID = POST_ID,
    AGE = POST_AGE,
    GENDER = POST_GENDER,
    DESIGN = POST_DESIGN,
    DATA = POST_DATA,
    POLITICS = POST_POLITICS,
    ALIGN = POST_ALIGN,
    INTENT = POST_INTENT,
    TRUST = POST_TRUST,
    BEAUTY = POST_BEAUTY,
    ID_CONF = POST_ID_CONF,
    AGE_CONF = POST_AGE_CONF,
    GENDER_CONF = POST_GENDER_CONF
  ) %>% 
  #get rid of magicwizard transform for raw z value
  mutate(
    DESIGN_z = as.numeric(POST_DESIGN_z),
    DATA_z = as.numeric(POST_DATA_z),
    POLITICS_z = as.numeric(POST_POLITICS_z),
    ALIGN_z = as.numeric(POST_ALIGN_z),
    TRUST_z = as.numeric(POST_TRUST_z),
    BEAUTY_z = as.numeric(POST_BEAUTY_z),
    INTENT_z = as.numeric(POST_INTENT_z),
    ID_CONF_z = as.numeric(POST_ID_CONF_z),
    AGE_CONF_z = as.numeric(POST_AGE_CONF_z),
    GENDER_CONF_z = as.numeric(POST_GENDER_CONF_z)
    ) %>% 
  # drop old columns datawizard columns
  select(-c(POST_DESIGN_z: POST_INTENT_z)) %>%
  select(-c(POST_ID_CONF_z, POST_AGE_CONF_z, POST_GENDER_CONF_z)) %>% 
  mutate(TIME="POST") %>% 
  droplevels()



  
## JOIN STUDIES
df <- rbind(df1, df3_PRE, df3_POST)

## PRINT 
print("PARTICIPANTS BY STUDY AND STIMULUS")
addmargins(table(df$Study, df$STIMULUS))

## WRITE FILE
saveRDS(df, file = "data/output/COMBINED/df_graphs_ALL.rds")

df_graphs_all <- df
## CLEANUP
rm(df, df1, df3)

```

## JOIN QUESTION LONG FILES
(useful for graphs and models that need all qs in long format, not just the sd questions)
```{r join-longqus}

## KEEP TIME COLUMN FOR GRAPHING

## WRANGLE STUDIES 1_3, 
df1 <- df_questions_long_s12 %>% 
  # Retain only cols needed for modelling
  select(PID, Assigned.Block, Study, Sample, D_politicsSocial, D_politicsFiscal, 
         STIMULUS_CATEGORY, STIMULUS, BLOCK, QUESTION, value) %>% 
  # Filter out questions dropped from Study 3
  filter(QUESTION %nin% c("MAKER_ARGUE","MAKER_SELF","CHART_LIKE","CHART_TRUST")) %>% 
  # Rename questions to match Study 3
  mutate(
    QUESTION = str_replace(QUESTION, "MAKER_CONF","ID_CONF"), #handle maker conf
    QUESTION = str_replace(QUESTION, "MAKER_",""), #strip MAKER_ prefix
    QUESTION = str_replace(QUESTION, "CHART_",""), #strip CHART_ prefix
    QUESTION = str_replace(QUESTION, "POLITIC","POLITICS"), 
    TIME = "PRE" #INDICATES measuremeent was on obscured graph 
  ) %>% 
  # Drop levels
  droplevels()

## Study 3, RETAIN ONLY COLUMNS NECESSARY FOR MODELLING, and only TIME=PRE
df3 <- df_questions_long_s3 %>% 
  #Retail only cols needed for modelling
  select(PID, Assigned.Block, Study, Sample, D_politicsSocial, D_politicsFiscal, 
        STIMULUS_CATEGORY, STIMULUS, BLOCK, QUESTION, value, TIME) %>% 
  ######### OLD CODE 
  # # Filter out responses from the POST test
  # filter(TIME=="PRE") %>% 
  # # Drop time
  # select(-TIME) %>% 
  # Drop levels
  droplevels()

## JOIN STUDIES
df <- rbind(df1, df3) %>% 
  mutate(TIME=factor(TIME)) %>% 
  #drop questions we don't use
  filter(QUESTION %nin% c("DETAIL","EXPLAIN","LATENCY","TOOL_ID","TOOL_CONF","TOOL_DETAIL","TYPE","ACTION"))


## PRINT 
print("PARTICIPANT-QUESTSIONS BY STUDY AND STIMULUS AND time")
addmargins(table(df$QUESTION, df$STIMULUS, df$Study))
table(df$Study, df$STIMULUS, df$TIME)

## WRITE FILE
saveRDS(df, file = "data/output/COMBINED/df_questions_long_ALL.rds")

## CLEANUP
rm(df, df1, df3)

```



## JOIN SD QUESTION LONG Files

```{r join-sds}

## KEEP TIME COLUMN FOR GRAPHING

## WRANGLE STUDIES 1_3, 
df1 <- df_sd_questions_long_s12 %>% 
  # Retain only cols needed for modelling
  select(PID, Assigned.Block, Study, Sample, D_politicsSocial, D_politicsFiscal, 
         STIMULUS_CATEGORY, STIMULUS, BLOCK, QUESTION, value) %>% 
  # Filter out questions dropped from Study 3
  filter(QUESTION %nin% c("MAKER_ARGUE","MAKER_SELF","CHART_LIKE","CHART_TRUST")) %>% 
  # Rename questions to match Study 3
  mutate(
    QUESTION = str_replace(QUESTION, "MAKER_",""), #strip MAKER_ prefix
    QUESTION = str_replace(QUESTION, "CHART_",""), #strip CHART_ prefix
    QUESTION = str_replace(QUESTION, "POLITIC","POLITICS"), 
    QUESTION = factor(QUESTION, levels = ref_min_sd_questions), #set factor order
    TIME = "PRE" #INDICATES measuremeent was on obscured graph 
  ) %>% 
  # Drop levels
  droplevels()

## Study 3, RETAIN ONLY COLUMNS NECESSARY FOR MODELLING, and only TIME=PRE
df3 <- df_sd_questions_long_s3 %>% 
  #Retail only cols needed for modelling
  select(PID, Assigned.Block, Study, Sample, D_politicsSocial, D_politicsFiscal, 
        STIMULUS_CATEGORY, STIMULUS, BLOCK, QUESTION, value, TIME) %>% 
  ######### OLD CODE 
  # # Filter out responses from the POST test
  # filter(TIME=="PRE") %>% 
  # # Drop time
  # select(-TIME) %>% 
  # Drop levels
  droplevels()

## JOIN STUDIES
df <- rbind(df1, df3) %>% 
  mutate(TIME=factor(TIME))


## PRINT 
print("PARTICIPANT-QUESTSIONS BY STUDY AND STIMULUS AND time")
addmargins(table(df$QUESTION, df$STIMULUS, df$Study))
table(df$Study, df$STIMULUS, df$TIME)

## WRITE FILE
saveRDS(df, file = "data/output/COMBINED/df_sd_questions_long_ALL.rds")

## CLEANUP
rm(df, df1, df3)

```

## JOIN SD QUESTION LONG Files (Z-SCORED)

```{r join-sds-z}

## WRANGLE STUDIES 1_3, 
df1 <- df_sd_questions_long_s12_z %>% 
  # Retain only cols needed for modelling
  select(PID, Assigned.Block, Study, Sample, D_politicsSocial, D_politicsFiscal, 
         STIMULUS_CATEGORY, STIMULUS, BLOCK, QUESTION, value) %>% 
  # Filter out questions dropped from Study 3
  filter(QUESTION %nin% c("MAKER_ARGUE_z","MAKER_SELF_z","CHART_LIKE_z","CHART_TRUST_z"))  %>% 
  # Rename questions to match Study 3
  mutate(
    QUESTION = str_replace(QUESTION, "MAKER_",""), #strip MAKER_ prefix
    QUESTION = str_replace(QUESTION, "CHART_",""), #strip CHART_ prefix
    QUESTION = str_replace(QUESTION, "POLITIC","POLITICS"), 
    QUESTION = factor(QUESTION, levels = ref_min_sd_questions_z), #set factor order
    TIME = "PRE"
    ) %>% 
  # Drop levels
  droplevels()

## STUDY 3, RETAIN ONLY COLUMNS NECESSARY FOR MODELLING, and only TIME=PRE
df3 <- df_sd_questions_long_s3_z %>% 
  #Retail only cols needed for modelling
  select(PID, Assigned.Block, Study, Sample, D_politicsSocial, D_politicsFiscal, 
        STIMULUS_CATEGORY, STIMULUS, BLOCK, QUESTION, value, TIME) %>% 
  # # Filter out responses from the POST test
  # filter(TIME=="PRE") %>% 
  # # Drop time
  # select(-TIME) %>% 
  # Drop levels
  droplevels()

## JOIN STUDIES
df <- rbind(df1, df3) %>% 
  mutate(TIME=factor(TIME))

## PRINT 
print("PARTICIPANT-QUESTSIONS BY STUDY AND STIMULUS")
addmargins(table(df$QUESTION, df$STIMULUS, df$Study))
table(df$Study, df$STIMULUS, df$TIME)


## WRITE FILE
saveRDS(df, file = "data/output/COMBINED/df_sd_questions_long_ALL_z.rds")

#cleanup
rm(df1, df3, df)

```

# PROFILING

## PARTICIPANT-LEVEL

```{r summarytools-profile-participants, results='asis'}


dfSummary(df_participants_all,  #%>% select(PID, duration.min, Distribution, Assigned.Block, PLATFORM, 
                      #       contains("D_"), Prolific.Name:Study, contains("SCREEN_")), 
          headings = TRUE,
          plain.ascii  = FALSE,
          style        = 'grid',
          graph.magnif = 0.85,
          varnumbers = FALSE,
          valid.col    = FALSE,
          tmp.img.dir  = "/tmp")


```

## GRAPH/TRIAL-LEVEL

```{r summarytools-profile-graphs, results='asis'}


dfSummary(df_graphs_all,  # %>% select(PID, duration.min, Distribution, Assigned.Block, PLATFORM,
                      #       contains("D_"), Prolific.Name:Study, contains("SCREEN_")), 
          headings = TRUE,
          plain.ascii  = FALSE,
          style        = 'grid',
          graph.magnif = 0.85,
          varnumbers = FALSE,
          valid.col    = FALSE,
          tmp.img.dir  = "/tmp")


```
